+++ include/MenuFolder.h
@@ -30,11 +30,8 @@ public:
     virtual int initialize();
 
     virtual void display();
-
     virtual void handleKeyboard(KeyboardButton key, bool pressed);
-
     virtual void handleMouseClick(unsigned int x, unsigned int y, KeyboardButton button, bool released);
-
     virtual void handleMouseScroll(int xrel, int yrel);
 
 private: