+++ src/Game.cpp
@@ -30,14 +30,19 @@ std::map<int, int> gMapTex2Bump;
 #endif
 
 Game::Game() {
+    zPos = 0;
     mLoaded = false;
     mLara = -1;
     mTextureStart = 0;
     mTextureOffset = 0;
+
+    UI::addWindow(this);
 }
 
 Game::~Game() {
     destroy();
+
+    UI::removeWindow(this);
 }
 
 unsigned int Game::getTextureStart() {
@@ -55,6 +60,10 @@ int Game::initialize() {
     mTextureStart = getTextureManager().getTextureCount();
 
     return 0;
+}
+
+void Game::display() {
+    getRender().display();
 }
 
 void Game::destroy() {