+++ src/Menu.cpp
@@ -11,14 +11,6 @@
 #include "Window.h"
 #include "Menu.h"
 
-void Menu::setVisible(bool visible) {
-    mVisible = visible;
-}
-
-bool Menu::isVisible() {
-    return mVisible;
-}
-
 void Menu::showDialog(std::string msg, std::string btn1, std::string btn2,
         std::function<int (bool state)> callback) {
     // Only show one dialog at a time
@@ -94,7 +86,7 @@ bool Menu::handleMouseScrollDialog(int xrel, int yrel) {
 
 void Menu::displayDialog() {
     if (dialogText.length() > 0) {
-        unsigned int wMax = ((unsigned int)(getWindow().getWidth() * 0.66f));
+        unsigned int wMax = ((unsigned int)(::getWindow().getWidth() * 0.66f));
 
         unsigned int w0 = getFont().widthText(1.0f, dialogText) + 20;
         if (w0 > wMax)
@@ -146,8 +138,8 @@ void Menu::displayDialog() {
             hOverlay = h0 + h1;
         }
 
-        unsigned int xOverlay = (getWindow().getWidth() - wOverlay) / 2;
-        unsigned int yOverlay = (getWindow().getHeight() - hOverlay) / 2;
+        unsigned int xOverlay = (::getWindow().getWidth() - wOverlay) / 2;
+        unsigned int yOverlay = (::getWindow().getHeight() - hOverlay) / 2;
 
         glColor4f(0.0f, 0.0f, 0.0f, 0.75f);
         glDisable(GL_TEXTURE_2D);
@@ -162,13 +154,13 @@ void Menu::displayDialog() {
                 getFont().drawTextWrapped(xOverlay + 10 + w1, yOverlay + 10 + h0,
                     1.0f, dialogState ? RED : BLUE, w2, dialogButton2);
             } else {
-                getFont().drawTextWrapped((getWindow().getWidth() - w1) / 2,
+                getFont().drawTextWrapped((::getWindow().getWidth() - w1) / 2,
                     yOverlay + 10 + h0, 1.0f, dialogState ? BLUE : RED, w1, dialogButton1);
-                getFont().drawTextWrapped((getWindow().getWidth() - w2) / 2,
+                getFont().drawTextWrapped((::getWindow().getWidth() - w2) / 2,
                     yOverlay + 10 + h0 + h1, 1.0f, dialogState ? RED : BLUE, w2, dialogButton2);
             }
         } else {
-            getFont().drawTextWrapped((getWindow().getWidth() - w1) / 2,
+            getFont().drawTextWrapped((::getWindow().getWidth() - w1) / 2,
                     yOverlay + 10 + h0, 1.0f, RED, w1, dialogButton1);
         }
     }